fix(admin): render video detail as page
